Split `lib.rs` into multiple files
This makes the code slightly more readable.
